# docs/swagger.yaml
basePath: /
definitions:
  structures.Password:
    properties:
      password:
        type: string
    type: object
  structures.StatusBadRequest:
    properties:
      error:
        example: invalid request
        type: string
    type: object
  structures.StatusInternalServerError:
    properties:
      error:
        example: internal server error
        type: string
    type: object
  structures.StatusNotFound:
    properties:
      message:
        example: 'Error: not found'
        type: string
    type: object
  structures.StatusOK:
    properties:
      status:
        example: ok
        type: string
    type: object
  structures.StatusUnauthorized:
    properties:
      error:
        example: unauthorized
        type: string
    type: object
  structures.Task:
    properties:
      comment:
        type: string
      date:
        type: string
      id:
        type: string
      repeat:
        type: string
      title:
        type: string
    type: object
host: localhost:7540
info:
  contact: {}
  description: API Server for TodoList Application
  title: TodoList API
  version: "1.0"
paths:
  /api/nextdate:
    get:
      consumes:
      - application/json
      description: get next date
      operationId: next-date
      parameters:
      - description: now
        in: query
        name: now
        required: true
        type: string
      - description: date
        in: query
        name: date
        required: true
        type: string
      - description: repeat
        in: query
        name: repeat
        required: true
        type: string
      produces:
      - application/json
      responses:
        "200":
          description: OK
          schema:
            $ref: '#/definitions/structures.StatusOK'
        "400":
          description: Bad Request
          schema:
            $ref: '#/definitions/structures.StatusBadRequest'
        "401":
          description: Unauthorized
          schema:
            $ref: '#/definitions/structures.StatusUnauthorized'
        "404":
          description: Not Found
          schema:
            $ref: '#/definitions/structures.StatusNotFound'
        "500":
          description: Internal Server Error
          schema:
            $ref: '#/definitions/structures.StatusInternalServerError'
      summary: Next date
      tags:
      - next-date
  /api/signin:
    post:
      consumes:
      - application/json
      description: login
      operationId: login
      parameters:
      - description: password
        in: body
        name: password
        required: true
        schema:
          $ref: '#/definitions/structures.Password'
      produces:
      - application/json
      responses:
        "200":
          description: OK
          schema:
            $ref: '#/definitions/structures.StatusOK'
        "400":
          description: Bad Request
          schema:
            $ref: '#/definitions/structures.StatusBadRequest'
        "401":
          description: Unauthorized
          schema:
            $ref: '#/definitions/structures.StatusUnauthorized'
        "404":
          description: Not Found
          schema:
            $ref: '#/definitions/structures.StatusNotFound'
        "500":
          description: Internal Server Error
          schema:
            $ref: '#/definitions/structures.StatusInternalServerError'
      summary: Login
      tags:
      - login
  /api/task:
    delete:
      consumes:
      - application/json
      description: delete task
      operationId: delete-task
      parameters:
      - description: id
        in: query
        name: id
        required: true
        type: string
      produces:
      - application/json
      responses:
        "200":
          description: OK
          schema:
            $ref: '#/definitions/structures.StatusOK'
        "400":
          description: Bad Request
          schema:
            $ref: '#/definitions/structures.StatusBadRequest'
        "401":
          description: Unauthorized
          schema:
            $ref: '#/definitions/structures.StatusUnauthorized'
        "404":
          description: Not Found
          schema:
            $ref: '#/definitions/structures.StatusNotFound'
        "500":
          description: Internal Server Error
          schema:
            $ref: '#/definitions/structures.StatusInternalServerError'
      summary: Delete task
      tags:
      - tasks
    get:
      consumes:
      - application/json
      description: get task
      operationId: get-task
      parameters:
      - description: id
        in: query
        name: id
        required: true
        type: string
      produces:
      - application/json
      responses:
        "200":
          description: OK
          schema:
            $ref: '#/definitions/structures.StatusOK'
        "400":
          description: Bad Request
          schema:
            $ref: '#/definitions/structures.StatusBadRequest'
        "401":
          description: Unauthorized
          schema:
            $ref: '#/definitions/structures.StatusUnauthorized'
        "404":
          description: Not Found
          schema:
            $ref: '#/definitions/structures.StatusNotFound'
        "500":
          description: Internal Server Error
          schema:
            $ref: '#/definitions/structures.StatusInternalServerError'
      summary: Get task
      tags:
      - tasks
    post:
      consumes:
      - application/json
      description: create task
      operationId: create-task
      parameters:
      - description: task
        in: body
        name: task
        required: true
        schema:
          $ref: '#/definitions/structures.Task'
      produces:
      - application/json
      responses:
        "200":
          description: OK
          schema:
            $ref: '#/definitions/structures.StatusOK'
        "400":
          description: Bad Request
          schema:
            $ref: '#/definitions/structures.StatusBadRequest'
        "401":
          description: Unauthorized
          schema:
            $ref: '#/definitions/structures.StatusUnauthorized'
        "404":
          description: Not Found
          schema:
            $ref: '#/definitions/structures.StatusNotFound'
        "500":
          description: Internal Server Error
          schema:
            $ref: '#/definitions/structures.StatusInternalServerError'
      summary: Create task
      tags:
      - tasks
    put:
      consumes:
      - application/json
      description: update task
      operationId: update-task
      parameters:
      - description: task
        in: body
        name: task
        required: true
        schema:
          $ref: '#/definitions/structures.Task'
      produces:
      - application/json
      responses:
        "200":
          description: OK
          schema:
            $ref: '#/definitions/structures.StatusOK'
        "400":
          description: Bad Request
          schema:
            $ref: '#/definitions/structures.StatusBadRequest'
        "401":
          description: Unauthorized
          schema:
            $ref: '#/definitions/structures.StatusUnauthorized'
        "404":
          description: Not Found
          schema:
            $ref: '#/definitions/structures.StatusNotFound'
        "500":
          description: Internal Server Error
          schema:
            $ref: '#/definitions/structures.StatusInternalServerError'
      summary: Update task
      tags:
      - tasks
  /api/task/done:
    post:
      consumes:
      - application/json
      description: done task
      operationId: done-task
      parameters:
      - description: id
        in: query
        name: id
        required: true
        type: string
      produces:
      - application/json
      responses:
        "200":
          description: OK
          schema:
            $ref: '#/definitions/structures.StatusOK'
        "400":
          description: Bad Request
          schema:
            $ref: '#/definitions/structures.StatusBadRequest'
        "401":
          description: Unauthorized
          schema:
            $ref: '#/definitions/structures.StatusUnauthorized'
        "404":
          description: Not Found
          schema:
            $ref: '#/definitions/structures.StatusNotFound'
        "500":
          description: Internal Server Error
          schema:
            $ref: '#/definitions/structures.StatusInternalServerError'
      summary: Done task
      tags:
      - tasks
  /api/tasks:
    get:
      consumes:
      - application/json
      description: get all tasks
      operationId: get-tasks
      parameters:
      - description: search
        in: query
        name: search
        type: string
      produces:
      - application/json
      responses:
        "200":
          description: OK
          schema:
            items:
              $ref: '#/definitions/structures.StatusOK'
            type: array
        "400":
          description: Bad Request
          schema:
            $ref: '#/definitions/structures.StatusBadRequest'
        "401":
          description: Unauthorized
          schema:
            $ref: '#/definitions/structures.StatusUnauthorized'
        "404":
          description: Not Found
          schema:
            $ref: '#/definitions/structures.StatusNotFound'
        "500":
          description: Internal Server Error
          schema:
            $ref: '#/definitions/structures.StatusInternalServerError'
      summary: Get tasks
      tags:
      - tasks
swagger: "2.0"